product add update and completed, product details,cart,checkout page update

+++ src/pages/Cart/CartItems/CartItem.jsx
@@ -1,39 +1,24 @@
-import { useEffect, useState } from "react";
+import { MdOutlineDelete } from "react-icons/md";
 import { FiMinusCircle, FiPlusCircle } from "react-icons/fi";
 import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
 import Swal from "sweetalert2";
 import { changeQuantity, removeFromCart } from "../../../Redux/cart/cartSlice";
-import { useGetProductBySlugQuery } from "../../../Redux/product/productApi";
 
 export default function CartItem({ product }) {
   const dispatch = useDispatch();
-  const [stock, setStock] = useState(null);
-
-  const { slug, image, title, discount, price, quantity, size, color } =
+  const { slug, image, title, discount, price, quantity, size, color, stock } =
     product;
-  const { data } = useGetProductBySlugQuery(slug);
 
-  useEffect(() => {
-    const varients = data?.data?.varients;
-
-    const selectedColorVarient = varients?.find(
-      (varient) => varient.color === color
-    );
-
-    const selectedSizeVarient = selectedColorVarient?.info?.find(
-      (varient) => varient.size === size
-    );
-
-    setStock(selectedSizeVarient?.quantity);
-  }, [color, data?.data?.varients, size]);
+  console.log(product);
 
   const discountPrice = parseInt(price - (price * discount) / 100);
   const total =
     parseInt(discount >= 1 ? discountPrice : price) * parseInt(quantity);
 
   const handelDeleteCartItem = (data) => {
-    dispatch(removeFromCart(data));
+    const isConfirm = window.confirm("Are you sure delete this item?");
+    if (isConfirm) dispatch(removeFromCart(data));
   };
 
   const handelIncreaseQuantity = () => {
@@ -59,15 +44,16 @@ export default function CartItem({ product }) {
   return (
     <tr>
       <td className="p-3">
-        <div className="w-max flex gap-2 items-center">
+        <div className="flex gap-2 items-center">
           <img
             src={`${import.meta.env.VITE_BACKEND_URL}/products/${image}`}
             alt={title}
             className="w-10 h-10 rounded-lg"
           />
           <Link to={`/product/${slug}`} className="leading-4">
-            <h3 className="text-[17px] text-neutral">
-              {title.length > 30 ? `${title.slice(0, 30)}...` : title}
+            <h3 className="text-neutral">
+              {/* {title.length > 30 ? `${title.slice(0, 30)}...` : title} */}
+              {title}
             </h3>
             <p className="text-neutral-content">
               {size && <small>size: {size}</small>}{" "}
@@ -78,34 +64,34 @@ export default function CartItem({ product }) {
       </td>
 
       <td className="px-6 py-2 font-medium">
-        <p className="w-max">
+        <p>
           {discount >= 1 ? (
             <>
               <span>৳{discountPrice}</span>
               <del className="text-xs text-neutral/80 pl-1">৳{price}</del>
             </>
           ) : (
-            <span>৳ {price}</span>
+            <span>৳{price}</span>
           )}
         </p>
       </td>
 
       <td className="px-6 py-3">
-        <div className="w-max flex items-center gap-3">
+        <div className="flex items-center gap-2">
           <button
             onClick={handelDecreaseQuantity}
-            className="text-2xl hover:text-neutral duration-200"
+            className="text-lg text-neutral-content hover:text-neutral duration-200"
           >
             <FiMinusCircle />
           </button>
           <div>
-            <p className="w-14 border-2 border-neutral/80 text-neutral font-medium text-center rounded-lg py-px">
+            <p className="w-max  text-neutral text-center rounded-lg py-px">
               {quantity}
             </p>
           </div>
           <button
             onClick={handelIncreaseQuantity}
-            className="text-2xl hover:text-neutral duration-200"
+            className="text-lg text-neutral-content hover:text-neutral duration-200"
           >
             <FiPlusCircle />
           </button>
@@ -113,17 +99,17 @@ export default function CartItem({ product }) {
       </td>
 
       <td className="px-6 py-3 font-medium">
-        <p className="w-max">
-          ৳<span> {total}</span>
+        <p>
+          ৳<span>{total}</span>
         </p>
       </td>
 
       <td className="px-6 py-3">
         <button
           onClick={() => handelDeleteCartItem(product)}
-          className="font-medium text-sm text-red-600 hover:underline"
+          className="text-red-600 hover:underline text-xl"
         >
-          Remove
+          <MdOutlineDelete />
         </button>
       </td>
     </tr>
